requirements.txt
google-auth==2.34.0
gspread==6.1.2
gspread-dataframe==4.0.0
numpy==2.1.1
pandas==2.2.2
PyMySQL==1.1.1
python-dotenv==1.0.1